Refactor status feature
