+++ src/features/Status/StatusPage.tsx
@@ -1,7 +1,7 @@
 import { Flex } from '@rebass/grid';
 import * as React from 'react';
 import { BackgroundImage, Button, H1, Paragraph } from '../../shared/Elements';
-import SideBarComponent from './Sidebar';
+import SideBarComponent from '../Dashboard/Sidebar';
 
 import Background from '../../assets/images/statuspage-background.svg';
